Refactor navigation
Implemented routing for the login and recording pages.
Broke up the views into their own files.
