+++ lib/components.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+final offTopTitle = AppBar(
+  title: Text(
+    'OFF-TOP',
+    style: TextStyle(
+        color: Colors.black45, fontSize: 30, fontWeight: FontWeight.w600),
+    textAlign: TextAlign.center,
+  ),
+  centerTitle: true,
+  backgroundColor: Colors.white,
+);
