+++ lib/router.dart
@@ -3,6 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:off_top_mobile/main.dart';
 import 'package:off_top_mobile/routing_constants.dart';
 import 'package:off_top_mobile/websocket.dart';
+import 'package:off_top_mobile/login.dart';
+import 'package:off_top_mobile/recording.dart';
 
 Route<dynamic> generateRoute(RouteSettings settings) {
   switch (settings.name) {
@@ -10,7 +12,11 @@ Route<dynamic> generateRoute(RouteSettings settings) {
       return MaterialPageRoute(builder: (_) => MyApp());
     case WebsocketRoute:
       return MaterialPageRoute(builder: (_) => WebsocketPage());
+    case LoginRoute:
+      return MaterialPageRoute(builder: (_) => LoginPage());
+    case RecordingRoute:
+      return MaterialPageRoute(builder: (_) => RecordingPage());
     default:
-      return MaterialPageRoute(builder: (_) => MyApp());
+      return MaterialPageRoute(builder: (_) => LoginPage());
   }
 }
