+++ lib/routing_constants.dart
@@ -1,2 +1,4 @@
-const String HomeRoute = '/home';
+const String RecordingRoute = '/recording';
 const String WebsocketRoute = '/websocket';
+const String LoginRoute = '/login';
+const String HomeRoute = '/home';
